Some changes to the controllers/services and html
I have added a migration for a menu item within the dashboard, changes to the controller and service provider/services, as well as changes to the html to be more similar to the dashboard.

The changes to the files was mostly to make the access token function properly without erroring. it should (tm) be fairly safe now.

all in all, provided the facebook side works properly and the changes to ENV are done to make it connect to the right facebook app it SHOULD work (but i can't test that right now.)